refactor: text - separate rune and byte string header files
