+++ src/text/rstring.c
@@ -10,8 +10,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "text.h"
+#include "rstring.h"
+#include "runes.h"
 
+// utf8_length returns the number of utf-8 characters in a string.
 static size_t utf8_length(const char* str) {
     size_t length = 0;
 
@@ -38,33 +40,40 @@ static size_t utf8_length(const char* str) {
     return length;
 }
 
+// string_new creates an empty string.
 static RuneString string_new(void) {
     RuneString str = {.runes = NULL, .length = 0, .size = 0, .owning = true};
     return str;
 }
 
+// string_from_runes creates a new string from an array of utf-8 characters.
+// `owning` indicates whether the string owns the array and should free the memory when destroyed.
 static RuneString string_from_runes(const int32_t* const runes, size_t length, bool owning) {
     RuneString str = {
         .runes = runes, .length = length, .size = length * sizeof(int32_t), .owning = owning};
     return str;
 }
 
+// string_from_cstring creates a new string from a zero-terminated C string.
 static RuneString string_from_cstring(const char* const utf8str) {
     size_t length = utf8_length(utf8str);
     int32_t* runes = length > 0 ? runes_from_cstring(utf8str, length) : NULL;
     return string_from_runes(runes, length, true);
 }
 
+// string_to_cstring converts the string to a zero-terminated C string.
 static char* string_to_cstring(RuneString str) {
     return runes_to_cstring(str.runes, str.length);
 }
 
+// string_free destroys the string, freeing resources if necessary.
 static void string_free(RuneString str) {
     if (str.owning && str.runes != NULL) {
         free((void*)str.runes);
     }
 }
 
+// string_at returns a character by its index in the string.
 static int32_t string_at(RuneString str, size_t idx) {
     if (str.length == 0) {
         return 0;
@@ -75,6 +84,9 @@ static int32_t string_at(RuneString str, size_t idx) {
     return str.runes[idx];
 }
 
+// string_slice returns a slice of the string,
+// from the `start` index (inclusive) to the `end` index (non-inclusive).
+// Negative `start` and `end` values count from the end of the string.
 static RuneString string_slice(RuneString str, int start, int end) {
     if (str.length == 0) {
         return string_new();
@@ -97,6 +109,8 @@ static RuneString string_slice(RuneString str, int start, int end) {
     return slice;
 }
 
+// string_substring returns a substring of `length` characters,
+// starting from the `start` index.
 static RuneString string_substring(RuneString str, size_t start, size_t length) {
     if (length > str.length - start) {
         length = str.length - start;
@@ -104,6 +118,8 @@ static RuneString string_substring(RuneString str, size_t start, size_t length) 
     return string_slice(str, start, start + length);
 }
 
+// string_contains_after checks if the other string is a substring of the original string,
+// starting at the `start` index.
 static bool string_contains_after(RuneString str, RuneString other, size_t start) {
     if (start + other.length > str.length) {
         return false;
@@ -116,6 +132,8 @@ static bool string_contains_after(RuneString str, RuneString other, size_t start
     return true;
 }
 
+// string_index_char returns the first index of the character in the string
+// after the `start` index, inclusive.
 static int string_index_char(RuneString str, int32_t rune, size_t start) {
     for (size_t idx = start; idx < str.length; idx++) {
         if (str.runes[idx] == rune) {
@@ -125,6 +143,8 @@ static int string_index_char(RuneString str, int32_t rune, size_t start) {
     return -1;
 }
 
+// string_index_char returns the last index of the character in the string
+// before the `end` index, inclusive.
 static int string_last_index_char(RuneString str, int32_t rune, size_t end) {
     if (end >= str.length) {
         return -1;
@@ -137,6 +157,8 @@ static int string_last_index_char(RuneString str, int32_t rune, size_t end) {
     return -1;
 }
 
+// string_index_after returns the index of the substring in the original string
+// after the `start` index, inclusive.
 static int string_index_after(RuneString str, RuneString other, size_t start) {
     if (other.length == 0) {
         return start;
@@ -159,10 +181,12 @@ static int string_index_after(RuneString str, RuneString other, size_t start) {
     return -1;
 }
 
+// string_index returns the first index of the substring in the original string.
 static int string_index(RuneString str, RuneString other) {
     return string_index_after(str, other, 0);
 }
 
+// string_last_index returns the last index of the substring in the original string.
 static int string_last_index(RuneString str, RuneString other) {
     if (other.length == 0) {
         return str.length - 1;
@@ -186,6 +210,7 @@ static int string_last_index(RuneString str, RuneString other) {
     return -1;
 }
 
+// string_reverse returns the reversed string.
 static RuneString string_reverse(RuneString str) {
     int32_t* runes = (int32_t*)str.runes;
     for (size_t i = 0; i < str.length / 2; i++) {
@@ -197,8 +222,8 @@ static RuneString string_reverse(RuneString str) {
     return res;
 }
 
-// return string_from_runes(str.runes, str.length, false);
-
+// string_pad_left pads the string to the specified length by prepending `fill` characters.
+// If the string is already longer than the specified length, it is truncated on the right.
 RuneString string_pad_left(RuneString str, size_t length, RuneString fill) {
     if (str.length >= length) {
         // If the string is already longer than length, return a truncated version of the string
@@ -233,6 +258,8 @@ RuneString string_pad_left(RuneString str, size_t length, RuneString fill) {
     return new_str;
 }
 
+// string_pad_right pads the string to the specified length by appending `fill` characters.
+// If the string is already longer than the specified length, it is truncated on the right.
 RuneString string_pad_right(RuneString str, size_t length, RuneString fill) {
     if (str.length >= length) {
         // If the string is already longer than length, return a truncated version of the string
@@ -267,6 +294,7 @@ RuneString string_pad_right(RuneString str, size_t length, RuneString fill) {
     return new_str;
 }
 
+// string_print prints the string to stdout.
 static void string_print(RuneString str) {
     if (str.length == 0) {
         printf("'' (len=0)\n");
